Rewards

+++ src/Shoppite.Core/Entities/Reward_Point_Log.cs
@@ -9,7 +9,7 @@ namespace Shoppite.Core.Entities
         public int Id { get; set; }
         public int? UserId { get; set; }
         public int? OrgId { get; set; }
-        public int Reward_points { get; set; }
+        public decimal Reward_points { get; set; }
         public string Reward_type  { get; set; }
         public string Operation_type { get; set; }
         public DateTime? Date_created { get; set; }
